User Profile and Admin Profile option Removed
