+++ src/app/Components/Admin/block-user/block-user.component.html
@@ -1,4 +1,32 @@
-<app-nav-user></app-nav-user>
+<!-- Admin Nav Start -->
+<nav class="navbar navbar-expand-lg navbar navbar-dark bg-primary">
+        <a [routerLink]="['/AdminMenu']" class="navbar-brand">Cognizant Trainee</a>
+        <button type="button" class="navbar-toggler" data-toggle="collapse" data-target="#navbarCollapse2">
+            <span class="navbar-toggler-icon"></span>
+        </button>
+    
+        <div class="collapse navbar-collapse" id="navbarCollapse2">
+            <div class="navbar-nav">
+                <a routerLink="/AdminMenu" class="nav-item nav-link active">Home</a>
+                <!-- <a  class="nav-item nav-link">About</a>
+                <a  class="nav-item nav-link">Products</a> -->
+            </div>  
+            <form class="form-inline ml-auto">
+                    <div class="nav-item dropdown" style="margin-right: 25px">
+                            <a class="nav-link text-white dropdown-toggle" data-toggle="dropdown">My Account</a>
+                            <div class="dropdown-menu">
+                            <a class="dropdown-item">Notifications</a>
+                            <a class="dropdown-item" routerLink="/" routerLinkActive="true">Logout</a>
+                    </div>
+                    </div>
+                    <div>
+                      </div>
+                    <div></div>
+            </form>
+          
+        </div>
+    </nav>
+<!-- Admin Nav End -->
 
 <div class="jumbotron">
     <div class="container text-center">
