+++ src/app/Components/Nav-Bar/nav-user/nav-user.component.html
@@ -1,5 +1,5 @@
 <nav class="navbar navbar-expand-lg navbar navbar-dark bg-primary">
-    <a [routerLink]="['']" class="navbar-brand">Cognizant Trainee</a>
+    <a [routerLink]="['/UserMenu']" class="navbar-brand">Cognizant Trainee</a>
     <button type="button" class="navbar-toggler" data-toggle="collapse" data-target="#navbarCollapse2">
         <span class="navbar-toggler-icon"></span>
     </button>
@@ -16,7 +16,7 @@
         <form class="form-inline ml-auto">
 
           <div>
-            <a routerLink='/Profile' class="nav-item nav-link active" style="color: white">My Profile</a>
+            <a routerLink='/userprofile' class="nav-item nav-link active" style="color: white">My Profile</a>
           </div>
                 <div class="nav-item dropdown" style="margin-right: 25px">
                         <a class="nav-link text-white dropdown-toggle" data-toggle="dropdown">My Account</a>
