--- src/app/Components/profile/profile.component.html
@@ -1,90 +1,0 @@
-<!-- Nav Here -->
-
-<nav class="navbar navbar-expand-lg navbar navbar-dark bg-primary">
-        <a  class="navbar-brand" style="color: white">Cognizant Trainee</a>
-        <button type="button" class="navbar-toggler" data-toggle="collapse" data-target="#navbarCollapse2">
-            <span class="navbar-toggler-icon"></span>
-        </button>
-    
-        <div class="collapse navbar-collapse" id="navbarCollapse2">
-            <div class="navbar-nav">
-                <!-- <a routerLink="/TrainerMenu" class="nav-item nav-link active">Home</a> -->
-                <!-- <a  class="nav-item nav-link">About</a>
-                <a  class="nav-item nav-link">Products</a> -->
-            </div>
-          
-    
-            
-            <form class="form-inline ml-auto">
-    
-              <div>
-                <a routerLink='/Profile' class="nav-item nav-link active" style="color: white">My Profile</a>
-              </div>
-                    <div class="nav-item dropdown" style="margin-right: 25px">
-                            <a class="nav-link text-white dropdown-toggle" data-toggle="dropdown">My Account</a>
-                            <div class="dropdown-menu">
-                            <a class="dropdown-item">Notifications</a>
-                            <a class="dropdown-item" routerLink="/" routerLinkActive="true">Logout</a>
-                            </div>
-                    </div>
-    
-                    <div>
-                    
-                      </div>
-                    <div></div>
-            </form>
-          
-        </div>
-    </nav>
-    <!-- Nav End -->
-
-<div class="container-fluid" >
-    
-    <h3 class="text-center text-capitalize p-5" style="color:black;"> My Profile</h3>
-<div class="row" >
-  <div class="col-sm-4 text-center">
-      <img class="img-thumbnail m-3" alt="NA" src="assets/imgs/login.png" width="300" style="height: 300px;">
-      <a ><button type="button" class="btn btn-info btn-md m-4" style="color:black;">Edit Profile Picture</button></a>
-  </div>
-  <div class="col-sm-8">
-              <div class="modal-body">
-                  <table class="table table-striped  table-bordered" style="color:black;">
-                      <tbody>
-                          <tr>
-                              <td>Name</td>
-                              <td *ngIf="profileData">{{profileData.userName}}</td>
-                            </tr>
-
-                          <tr>
-                              <td>Email-id</td>
-                              <td *ngIf="profileData">{{profileData.email}}</td>
-                          </tr>
-
-                          <tr>
-                              <td>Mobile Number</td>
-                              <td *ngIf="profileData">{{profileData.contactNumber}}</td>
-                          </tr>
-
-                          <tr>
-                              <td>Specialization</td>
-                              <td *ngIf="profileData">{{profileData.TrainerTechnology}}</td>
-                          </tr>
-
-                          <tr>
-                              <td>LinkedIN ID</td>
-                              <td *ngIf="profileData">{{profileData.linkdinUrl}}</td>
-                           
-                          </tr>
-                          <tr>
-                              <td>Experience</td>
-                              <td *ngIf="profileData">{{profileData.yearOfExperience}}</td>
-                          </tr>
-
-                      </tbody>
-                  </table>
-              </div>
-          </div>
-        </div>
-    </div>
-
-<app-footer></app-footer>
