--- src/app/Models/UserDtl.ts
@@ -1,5 +1,0 @@
-export class UserDtl
-{
-    email:string;
-    password:string;
-}
